Work around project.build.resources being overridden
When specifying a project.build.resources section in a child pom.xml,
the parent's resources section is overridden -- not appended to, as this
developer expected. Therefore, we have to specify src/main/resources
explicitly if we want to keep it as a place to put resources into.

To be careful, we prefix it by ${basedir}/ just to make sure that we can
simply copy this code to aggregate projects whenever needed without
breaking things.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>